refactor: use singular name
