+++ pkg/provider/gitlab/gitlab_test.go
@@ -11,7 +11,7 @@ import (
 	"testing"
 
 	"github.com/Masterminds/semver"
-	"github.com/go-semantic-release/semantic-release/pkg/providers"
+	"github.com/go-semantic-release/semantic-release/pkg/provider"
 	"github.com/go-semantic-release/semantic-release/pkg/semrel"
 	"github.com/stretchr/testify/require"
 	"github.com/xanzy/go-gitlab"
@@ -185,6 +185,6 @@ func TestGitlabCreateRelease(t *testing.T) {
 	repo, ts := getNewGitlabTestRepo(t)
 	defer ts.Close()
 	newVersion := semver.MustParse("2.0.0")
-	err := repo.CreateRelease(&providers.RepositoryRelease{NewVersion: newVersion, SHA: "deadbeef"})
+	err := repo.CreateRelease(&provider.RepositoryRelease{NewVersion: newVersion, SHA: "deadbeef"})
 	require.NoError(t, err)
 }
