+++ pkg/provider/repository.go
@@ -1,4 +1,4 @@
-package providers
+package provider
 
 import (
 	"regexp"
